% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dammmdatagen-package.R
\docType{package}
\name{dammmdatagen-package}
\alias{dammmdatagen}
\alias{dammmdatagen-package}
\title{dammmdatagen: Marketing Mix Modeling Data Generator}
\description{
This package aims at being a general purpose example data generator for Marketing Mix Modeling Benchmark purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://doktormike.github.io/dammmdatagen/}
}

}
\keyword{internal}
