% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{generateCompetitorData}
\alias{generateCompetitorData}
\title{Generate competitor media spending as list prices data}
\usage{
generateCompetitorData(fromDate = Sys.Date() - 1 * 365, toDate = Sys.Date(),
  mynames = c("competitor_a", "competitor_b", "competitor_c"))
}
\arguments{
\item{fromDate}{the beginning of the time series}

\item{toDate}{the end of the time series}

\item{mynames}{the names to attach to the generated data}
}
\value{
a tibble with the generated data one column for each element in name
}
\description{
Generate competitor media spending as list prices data
}
\examples{
library(ggplot2)
library(dplyr)
generateCompetitorData(Sys.Date()-30, Sys.Date()) \%>\%
gather(competitor, spend, -date) \%>\%
ggplot(aes(y=spend, x=date, fill=competitor)) +
geom_bar(stat="identity", position = position_stack())
}
