% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{generateCovariatesData}
\alias{generateCovariatesData}
\title{Generate a marketing mix modeling covariates data set}
\usage{
generateCovariatesData(fromDate = Sys.Date() - 1 * 365,
  toDate = Sys.Date(), sector = "telekom",
  onlineInsertionNames = c("display", "facebook", "search_branded"),
  offlineInsertionNames = c("tv", "radio", "ooh", "print"),
  priceNames = c("price_product_a", "price_product_b",
  "price_product_c"), distributionNames = c("dist_product_a",
  "dist_product_b", "dist_product_c"), weatherNames = c("sunshine",
  "precipitation", "temperature"), competitorNames = c("competitor_a",
  "competitor_b", "competitor_c"), macroNames = c("cpi", "cci", "gdp"),
  eventNames = c("event_a", "event_b"))
}
\arguments{
\item{fromDate}{the beginning of the time series}

\item{toDate}{the end of the time series}

\item{sector}{the name of the sector to simulate (currently not used)}

\item{onlineInsertionNames}{the names of each online media insertion you wish to use}

\item{offlineInsertionNames}{the names of each offline media insertion you wish to use}

\item{priceNames}{the names of the different product prices}

\item{distributionNames}{the names of the different product distributions}

\item{weatherNames}{the names of the weather data measurements to use}

\item{competitorNames}{the names of each competitor considered}

\item{macroNames}{the names of the macroeconomical factors to simulate}

\item{eventNames}{the names of events to add}
}
\value{
a list of tibble containing each data mentioned
}
\description{
This function generates a marketing mix modeling covariates data set based on
the parameters given. It uses many different stochastical processes to
accomplish this and the dynamics behind them are not available to the user to
manipulate. This is supposed to generate all tha variables you use in a
marketing mix model except for the response variable i.e. the dependent
variable.
}
\examples{
a<-1
}
