% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{generateEventData}
\alias{generateEventData}
\title{Generate offline media data}
\usage{
generateEventData(
  fromDate = Sys.Date() - 1 * 365,
  toDate = Sys.Date(),
  mynames = c("event_a", "event_b"),
  freq = 0.01
)
}
\arguments{
\item{fromDate}{the beginning of the time series}

\item{toDate}{the end of the time series}

\item{mynames}{the names to attach to the generated data}

\item{freq}{the rate of incidence of events}
}
\value{
a tibble with the generated data one column for each element in name
}
\description{
Generate offline media data
}
\examples{
library(ggplot2)
library(dplyr)
library(tidyr)
generateEventData(Sys.Date() - 30, Sys.Date()) \%>\%
  gather(type, value, -date) \%>\%
  ggplot(aes(y = value, x = date, color = type)) +
  geom_line() +
  theme_minimal()
}
