% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{generateMacroData}
\alias{generateMacroData}
\title{Generate macro economical data}
\usage{
generateMacroData(fromDate = Sys.Date() - 1 * 365, toDate = Sys.Date(),
  mynames = c("cpi", "cci", "gdp"))
}
\arguments{
\item{fromDate}{the beginning of the time series}

\item{toDate}{the end of the time series}

\item{mynames}{the names to attach to the generated data}
}
\value{
a tibble with the generated data one column for each element in name
}
\description{
Generate macro economical data
}
\examples{
library(ggplot2)
library(dplyr)
library(tidyr)
generateMacroData(Sys.Date()-30, Sys.Date()) \%>\%
gather(type, measure, -date) \%>\%
ggplot(aes(y=measure, x=date, color=type)) +
geom_line() + theme_minimal()

ts.plot(arima.sim(list(order = c(1,1,0), ar = 0.7), n = 365))
}
