% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{generateOnlineData}
\alias{generateOnlineData}
\title{Generate online media data}
\usage{
generateOnlineData(fromDate = Sys.Date() - 1 * 365,
  toDate = Sys.Date(), mynames = c("display", "facebook",
  "search_branded"), avgcpm = 0.5, avgnet = 10000)
}
\arguments{
\item{fromDate}{the beginning of the time series}

\item{toDate}{the end of the time series}

\item{mynames}{the names to attach to the generated data}

\item{avgcpm}{the average Cost Per 1000 Impressions}

\item{avgnet}{the average net investment per insertion and day}
}
\value{
a tibble with the generated data one column for each element in name
}
\description{
Generate online media data
}
\examples{
library(ggplot2)
library(dplyr)
library(tidyr)
mydflist <- generateOnlineData(Sys.Date()-30, Sys.Date())
mydflist[["impression"]] \%>\%
gather(type, impression, -date) \%>\%
ggplot(aes(y=impression, x=date, color=type)) +
geom_line() + theme_minimal()
}
