% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{generatePriceData}
\alias{generatePriceData}
\title{Generate price data}
\usage{
generatePriceData(fromDate = Sys.Date() - 1 * 365, toDate = Sys.Date(),
  mynames = c("product_a", "product_b", "product_c"))
}
\arguments{
\item{fromDate}{the beginning of the time series}

\item{toDate}{the end of the time series}

\item{mynames}{the names to attach to the generated data}
}
\value{
a tibble with the generated data one column for each element in name
}
\description{
Generate price data
}
\examples{
library(ggplot2)
library(dplyr)
library(tidyr)
generatePriceData(Sys.Date()-30, Sys.Date()) \%>\%
gather(type, price, -date) \%>\%
ggplot(aes(y=price, x=date, color=type)) +
geom_line() + theme_minimal()
}
