% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{salesUnitsPerDay}
\alias{salesUnitsPerDay}
\title{Estimate units sold per day based on industry and population size of country.}
\usage{
salesUnitsPerDay(industry = "telecom", population = 5e+06)
}
\arguments{
\item{industry}{the industry to simulate which defaults to telecom}

\item{population}{the population size which defaults to 5 million people}
}
\value{
the number of expected sold units per average day
}
\description{
This utilizes heuristics to create a realistic estimation of how many units
a company can seel within an industry given the population size of that
country.
}
\examples{
library(ggplot2)
library(MASS)
avgsalesperday <- salesUnitsPerDay("telecom")
avgsales <- MASS::rnegbin(100, avgsalesperday[["UnitsPerDay"]], 1)
qplot(1:length(avgsales), avgsales) + geom_line()
}
