% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{generateOnlineData}
\alias{generateOnlineData}
\title{Generate price data}
\usage{
generateOnlineData(fromDate = Sys.Date() - 1 * 365, toDate = Sys.Date(),
  mynames = c("display", "facebook", "search_branded"))
}
\arguments{
\item{fromDate}{the beginning of the time series}

\item{toDate}{the end of the time series}

\item{mynames}{the names to attach to the generated data}
}
\value{
a tibble with the generated data one column for each element in name
}
\description{
Generate price data
}
\examples{
library(ggplot2)
library(dplyr)
library(tidyr)
generateOnlineData(Sys.Date()-30, Sys.Date()) \%>\%
gather(type, impression, -date) \%>\%
ggplot(aes(y=impression, x=date, color=type)) +
geom_line() + theme_minimal()
}
